+++ src/ImageInfo/Services/AdvancedBatchConverter.cs
@@ -268,7 +268,7 @@ namespace ImageInfo.Services
             {
                 using var image = new MagickImage(pngPath);
                 image.Format = MagickFormat.WebP;
-                image.Quality = quality;
+                image.Quality = (uint)quality;
                 image.Write(outPath);
 
                 sw.Stop();
@@ -311,7 +311,7 @@ namespace ImageInfo.Services
             {
                 using var image = new MagickImage(jpegPath);
                 image.Format = MagickFormat.WebP;
-                image.Quality = quality;
+                image.Quality = (uint)quality;
                 image.Write(outPath);
 
                 sw.Stop();
